Rename file name to addAIGeneratedContentsToThreads.ts
