+++ packages/server/graphql/mutations/helpers/handleCompletedStage.ts
@@ -10,7 +10,7 @@ import insertDiscussions from '../../../postgres/queries/insertDiscussions'
 import {AnyMeeting} from '../../../postgres/types/Meeting'
 import {DataLoaderWorker} from '../../graphql'
 import addDiscussionTopics from './addDiscussionTopics'
-import addAIGeneratedContentsToThreads from './addSummariesToThreads'
+import addAIGeneratedContentsToThreads from './addAIGeneratedContentsToThreads'
 import generateDiscussionSummary from './generateDiscussionSummary'
 import generateGroupSummaries from './generateGroupSummaries'
 import removeEmptyReflections from './removeEmptyReflections'
